+++ Fabric/src/main/java/de/markusbordihn/easymodelentities/gametest/HostEntityGameTest.java
@@ -60,4 +60,9 @@ public class HostEntityGameTest {
   public void blockEntityCanBePlacedAndInitialized(GameTestHelper helper) {
     HostEntityGameTestCases.blockEntityCanBePlacedAndInitialized(helper);
   }
+
+  @GameTest(template = Constants.MOD_ID + ":gametest.3x3x3")
+  public void attributesAreAppliedFromProfile(GameTestHelper helper) {
+    HostEntityGameTestCases.attributesAreAppliedFromProfile(helper);
+  }
 }